Gamestate and map
The API now has two new classes, GameState and Map. These are used to
maintain an updated state of the game world. The map object contains a
map data double array where each map cell is represented as a
javascript object with an j and k coordinate and a value denoting the
type of cell (grass, void etc).