Major changes in documentation, improvements to implementation
Documentation:
Added documentation for all code (methods, classes, etc). Documentation
is generated with Docco. Added Docco and grunt-docco to dev
dependencies in package.json. A grunt task for building the
documentation was also made and added to the watch task.

Structure:
Rebranded the repository/project as a Skyport API. The skyportai module
is now moved into an examples folder to illustrate how to use the api.

Code:
Added code to randomly move about the map.